% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose2.R
\name{choose2}
\alias{choose2}
\alias{combn2}
\title{Quicker Combinations IF Choosing 2}
\usage{
choose2(n = 2)

combn2(n = 2)
}
\arguments{
\item{n}{integer indicating the total number of cases from which to choose 2}
}
\description{
Calculate quicker combinations and choose output if only choosing 2
}
\note{
To make these functions fast, they do not check their arguments. Please
use these functions with caution, as incorrect arguments might have
unpredictable responses.
}
