% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{convert_TIRT}
\alias{convert_TIRT}
\title{Convert to thurstonianIRT quads}
\usage{
convert_TIRT(resp, n_item = 4)
}
\arguments{
\item{resp}{a list with:
(1) data.frame \code{item} indicating items, blocks, and dimensions, and
(2) data.frame \code{resp} indicating people, blocks, response numbers, and response sequences}

\item{n_item}{an integer indicating number of items per block.
currently only supports 4 items per block}
}
\value{
a data.frame following the structure of thurstonianIRT::triplets_long but with 4 items per block
}
\description{
Convert thirt responses with 4 items per block to thurstonianIRT data structure
}
\examples{
\dontrun{
set.seed(202106)

params <- simulate_thirt_params(n_person = 20,
                                n_item   = 4,
                                n_block  = 2,
                                n_dim    = 4)

resp   <- do.call(simulate_thirt_resp, params)

quads_long <- convert_TIRT(resp, n_item = 4)

}

}
