% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimate_thirt_params_mcmc}
\alias{estimate_thirt_params_mcmc}
\title{Estimate parameters with MCMC}
\usage{
estimate_thirt_params_mcmc(
  resp,
  items,
  control,
  initial_params = NULL,
  fixed_params = NULL
)
}
\arguments{
\item{fixed_params}{}
}
\value{

}
\description{
Estimate parameters with MCMC
}
\examples{

}
