% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{loglik_thirt}
\alias{loglik_thirt}
\title{Calculate log-likelihood}
\usage{
loglik_thirt(gamma, items, persons, resp)
}
\arguments{
\item{gamma}{a data.frame of length \verb{[total binary outcomes]} with two variables:
variable \code{pair} of the format \code{i-j} for item pair \code{ij},
variable \code{gamma} for threshold parameters.}

\item{items}{a data.frame of length \link{total items} with five variables:
variable \code{item} of the format \code{i} for item number \code{i},
variable \code{block} of the format \code{b} for block number \code{b},
variable \code{dim} of the format \code{d} for dimension number \code{d},
variable \code{lambda} for loadings,
variable \code{psisq} for uniqueness,
variable \code{dim} for dimensions.}

\item{persons}{a data.frame of length \verb{[number of people]} with variables:
variable \code{person} of the format \code{p} for person number \code{p},
variables named \code{theta_d} for dimension number \code{d}.}

\item{resp}{a data.frame of length \verb{[n_person x n_block]} with at least three first variables:
variable \code{person} of the format \code{p} for person number \code{p},
variable \code{block} of the format \code{b} for block number \code{b},
variable \code{resp} of the format \code{r} for response number \code{r}
which corresponds to mupp::find_permutation_index().}
}
\value{
a matrix of dimensions \verb{[n_person X n_block]} of log-likelihoods
of chosen responses for each person for each block.
}
\description{
Calculate log-likelihood
}
\examples{

\dontrun{
set.seed(202106)

params   <- simulate_thirt_params()

gamma    <- params$gamma
items    <- params$items
persons  <- params$persons
resp     <- do.call(simulate_thirt_resp, params)$resp

loglik_thirt(gamma = gamma, items = items, persons = persons, resp = resp)
}

}
