% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{prior}
\alias{prior}
\alias{d_thetas_prior}
\alias{d_lambda_prior}
\alias{d_psisq_prior}
\alias{d_gamma_prior}
\alias{r_thetas_prior}
\alias{r_lambda_prior}
\alias{r_psisq_prior}
\alias{r_gamma_prior}
\title{Prior distributions}
\usage{
d_thetas_prior(x, ...)

d_lambda_prior(x, direction = 1, ...)

d_psisq_prior(x, ...)

d_gamma_prior(x, ...)

r_thetas_prior(n, ...)

r_lambda_prior(n, ...)

r_psisq_prior(n, ...)

r_gamma_prior(n, ...)
}
\arguments{
\item{x}{vector of quantiles}

\item{...}{additional arguments}

\item{n}{number of observations for the simulation}
}
\description{
Prior distributions
}
