% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{sim_test}
\alias{sim_test}
\title{Simulation tests}
\usage{
sim_test(
  n_person,
  n_item,
  n_neg,
  n_block,
  n_dim,
  n_iter,
  n_burnin,
  step_size_sd,
  tirt = FALSE,
  seed = 20220414
)
}
\arguments{
\item{n_person}{integer indicating the number of people}

\item{n_item}{integer or vector of length \verb{[n_block]} indicating the number of items for each block}

\item{n_neg}{integer or vector of length \verb{[n_block]} indicating the number of negatively-keyed items for each block}

\item{n_block}{integer indicating the number of blocks}

\item{n_dim}{integer indicating the number of dimensions or traits}

\item{n_iter}{integer indicating the number of MCMC iterations}

\item{n_burnin}{integer indicating the number of MCMC burn-ins}

\item{step_size_sd}{double indicating the step size of new parameter generation for random-walk Metropolis step}

\item{tirt}{boolean indicating whether estimation results should be compared with the thurstonianIRT package}

\item{seed}{random seed}
}
\value{
save a timestamped RData file in the current working directory
}
\description{
Examine parameter recovery and performance using simulated data
}
\examples{
\dontrun{
sim_test(n_person = 10,
         n_item   = 4,
         n_neg    = 1,
         n_block  = 2,
         n_dim    = 4,
         n_iter   = 100,
         n_burnin = 1,
         step_size_sd = 0.1,
         tirt     = TRUE,
         seed     = 20220414)

}

}
