% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_thirt.R
\name{simulate_thirt_params}
\alias{simulate_thirt_params}
\title{Simulate parameters}
\usage{
simulate_thirt_params(n_person = 1, n_item = 2, n_block = 1, n_dim = 2)
}
\arguments{
\item{n_person}{integer indicating the number of people}

\item{n_item}{integer indicating the number of items per block}

\item{n_block}{integer indicating the number of blocks}

\item{n_dim}{integer indicating the number of dimensions or traits}
}
\value{
a list of parameters including gamma for item pairs,
individual item parameters, and person parameters
}
\description{
Generate parameters that conform to the Thurstonian IRT model.
}
