% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_thirt.R
\name{simulate_thirt_resp}
\alias{simulate_thirt_resp}
\title{Simulate THIRT responses}
\usage{
simulate_thirt_resp(gamma, items, persons)
}
\arguments{
\item{gamma}{a data.frame of length \verb{[total binary outcomes]} with two variables:
variable \code{pair} of the format \code{i-j} for item pair \code{ij}
variable \code{gamma} for threshold parameters}

\item{items}{a data.frame of length \verb{[total items]} with five variables:
variable \code{item} of the format \code{i} for item number \code{i},
variable \code{block} of the format \code{b} for block number \code{b},
variable \code{dim} of the format \code{d} for dimension number \code{d},
variable \code{lambda} for loadings,
variable \code{psisq} for uniqueness,
variable \code{dim} for dimensions}

\item{persons}{a data.frame of length \verb{[number of people]} with variables:
variable \code{person} of the format \code{p} for person number \code{p},
variables named \code{theta_d} for dimension number \code{d}.}
}
\value{
a list with:
(1) data.frame \code{items} indicating items, blocks, and dimensions and
(2) data.frame \code{resp} indicating people, blocks, response numbers, and response sequences.
}
\description{
Generate responses that can be used for the Thurstonian IRT model.
}
\examples{
\dontrun{
set.seed(202106)

params <- simulate_thirt_params()

do.call(simulate_thirt_resp, params)
}

}
