% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-thirt.R
\docType{package}
\name{thirt-package}
\alias{thirt}
\alias{thirt-package}
\title{thirt: Thurstonian FC-IRT Model}
\description{
Probability functions to describe the Thurstonian FC-IRT model,
simulate responses to the Thurstonian FC-IRT model, and estimate parameters
of the Thurstonian FC-IRT model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nguyenllpsych.github.io/thirt}
  \item \url{https://github.com/nguyenllpsych/thirt}
}

}
\author{
\strong{Maintainer}: Linh Nguyen \email{nguyenllpsych@gmail.com} (\href{https://orcid.org/0000-0002-1575-0395}{ORCID})

Authors:
\itemize{
  \item Steven Nydick \email{swnydick@gmail.com} (\href{https://orcid.org/0000-0002-2908-1188}{ORCID})
}

}
