% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{p_thirt}
\alias{p_thirt}
\title{Probability calculation}
\usage{
p_thirt(n_item, n_block, n_person, dict, gamma, lambda, theta, psisq)
}
\arguments{
\item{n_item}{number of items per block}

\item{n_block}{number of blocks}

\item{n_person}{number of respondents}

\item{dict}{a data.frame with a \code{dim} column identifying traits by number \verb{1, 2, 3...}
and an \code{item} column identifying item indicators for different traits by number \verb{1, 2, 3...}}

\item{gamma}{a data.frame with one variable of length \verb{[total binary outcomes]},
row names should be of the format \code{i-j} for item pair \code{ij}}

\item{lambda}{a data.frame with one variable of length \verb{[total items]},
row names should be of the format \code{i} for item \code{i}}

\item{theta}{a data.frame of dimension \verb{[number of people X number of dimensions]},
row names should be of the format \code{p} for person \code{p}}

\item{psisq}{a data.frame with one variable of length \verb{[total items]},
row names should be of the format \code{i} for item \code{i}}
}
\value{
a list of \verb{[permutation_list, probability]},
permutation_list is a matrix of dimension \verb{[number of permutations X number of items]},
probability is a matrix of dimension \verb{[person X permutation]} of probability for each response pattern
}
\description{
Calculate the probability of all potential response patterns for each respondent
}
\examples{
\dontrun{
set.seed(202106)

#####################
## Test Conditions ##
#####################

n_item    = 4
n_block   = 1
n_dim     = 4
n_person  = 200
block_size <- n_item * (n_item - 1) / 2

################
## Parameters ##
################

# thresholds - gamma
# vector of length [total binary outcomes]
# row names should be of the format "i-j" for item pair ij
gamma   <- data.frame(gamma = round(runif(n = block_size * n_block, min = -1,  max = 1), digits = 3))
row.names(gamma) <- c("1-2", "1-3", "1-4", "2-3", "2-4", "3-4")

# loadings - lambda
# vector of length [total items]
# row names should be of the format "i" for item i
lambda  <- data.frame(lambda = round(runif(n = n_item * n_block, min = .65, max = .95), digits = 3))

# uniqueness - psi^2
# vector of length [total items]
# row names should be of the format "i" for item i
psisq   <- data.frame(psisq = round(1 - lambda^2, digits = 3))

# theta - trait scores
# matrix of dimension [number of people X number of dimensions]
# row names should be of the format "p" for person p
theta   <- as.data.frame(matrix(nrow = n_person, ncol = n_dim))
for (dim in seq(n_dim)) {
  theta[dim] <- round(rnorm(n = n_person, mean = 0, sd = 1), digits = 3)
} # END for dim LOOP

# create a dictionary with a "dim" column identifying traits
#   and an "item" column identifying item indicators for different traits
dict <- data.frame(
  dim = c(seq(n_dim)),
  item = c(seq(n_item * n_block)))

###################
## Test Function ##
###################

# test function with fixed parameters
prob <- p_thirt(n_item, n_block, n_person, dict,
                gamma, lambda, theta, psisq)

prob$permutation_list
summary(prob$probability)
}

}
