% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_thirt.R
\name{simulate_thirt_params}
\alias{simulate_thirt_params}
\title{Simulate parameters}
\usage{
simulate_thirt_params(
  n_person = 2,
  n_item = c(2, 4),
  n_neg = 1,
  n_block = 2,
  n_dim = 4
)
}
\arguments{
\item{n_person}{integer indicating the number of people}

\item{n_item}{integer or vector of length \verb{[n_block]} indicating the number of items for each block}

\item{n_neg}{integer or vector of length \verb{[n_block]} indicating the number of negatively-keyed items for each block}

\item{n_block}{integer indicating the number of blocks}

\item{n_dim}{integer indicating the number of dimensions or traits}
}
\value{
a list of parameters including gamma for item pairs,
individual item parameters, and person parameters
}
\description{
Generate parameters that conform to the Thurstonian IRT model.
}
